backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
htmldocx==0.0.6
google-generativeai==0.8.3
supabase==2.9.1
python-jose[cryptography]==3.3.0
pydantic==2.9.2
httpx==0.27.2
